Use u64 for source location line and column

+++ atorsl/src/data/symbol.rs
@@ -5,8 +5,8 @@ use std::path::PathBuf;
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct SourceLoc {
     pub file: PathBuf,
-    pub line: u16,
-    pub col: u16,
+    pub line: u64,
+    pub col: u64,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
